feat: refactor hooks & add logs hook
